requirements.txt
torch>=1.5
torchvision
numpy>=1.16
pillow
skorch
scikit-learn
structlog
colorama
pandas
pyyaml
